refactor: extract run_aggregation_query into core module
